requirements.txt
<<<<<<< HEAD
# Core dependencies
=======
# 核心依赖
>>>>>>> e7862cd2291f87b9b6b2df0f04c4bd5cedbfdc39
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
<<<<<<< HEAD
xgboost>=1.7.0
lightgbm>=4.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# API service
=======

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0

# API 服务
>>>>>>> e7862cd2291f87b9b6b2df0f04c4bd5cedbfdc39
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
websockets>=11.0
<<<<<<< HEAD
python-multipart>=0.0.20
httpx>=0.28.0

# Configuration and logging
pyyaml>=6.0
python-dotenv>=1.0.0
structlog>=23.1.0
dpkt>=1.9.8

# Testing and quality
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.1.0
=======

# 配置管理
pyyaml>=6.0
python-dotenv>=1.0.0

# 日志
structlog>=23.1.0

# 测试
pytest>=7.4.0
pytest-cov>=4.1.0

# 代码质量
black>=23.0.0
flake8>=6.1.0

# 高级模型（可选，强烈推荐）
xgboost>=1.7.0
# lightgbm>=4.0.0

# 监控（可选）
# prometheus-client>=0.17.0
>>>>>>> e7862cd2291f87b9b6b2df0f04c4bd5cedbfdc39
